move route test files to appropriate folder
